added link to lecture 7 recording
note: this year's "introduction" to programming turned into an off-the-cuff hackathon, since all of the students in attendance already had programming experience.  to support students who were not able to attend the synchronous course meeting, i've also added a link to the recording of last year's analogous lecture, which provides a more standard introduction to programming.